+++ cmsafvis/R/render_plot.R
@@ -156,17 +156,22 @@ render_plot <- function(plot_rinstat,
 
       kml_toolbox <- raster::rasterToPolygons(ras)
 
-      check_package_dependency("plotKML", reason = "exporting KML files")
-
-      plotKML::plotKML(
-        kml_toolbox,
-        file = outfile,
-        kmz = FALSE,
-        open.kml = FALSE,
-        plot.labpt = FALSE,
-        overwrite = TRUE,
-        outline = 0
-      )
+      # check_package_dependency("plotKML", reason = "exporting KML files")
+      # 
+      # plotKML::plotKML(
+      #   kml_toolbox,
+      #   file = outfile,
+      #   kmz = FALSE,
+      #   open.kml = FALSE,
+      #   plot.labpt = FALSE,
+      #   overwrite = TRUE,
+      #   outline = 0
+      # )
+      
+      cat("Sorry, but the plotKML R-package was removed from CRAN 
+              and KML output is not possible at the moment.
+              We are working on a solution for the next update.","\n")
+      
     } else if (fileExtension == ".tif") {
       dta <- as.vector(visualizeDataTimestep)
       grd_dta <-